+++ CHANGELOG.md
@@ -8,9 +8,13 @@
 
 ## master branch (latest changes not released yet)
 
-- improved notebooks
+
+## 1.4.4 2020-10-14
+
+- improved notebooks (thanks to phidl tutorial)
 - added C and L components from phidl
 - print(component) returns more info (similar to phidl)
+- support new way of defining waveguides with pp.Path, pp.CrossSection and pp.path (thanks to phidl)
 
 ## 1.4.3 2020-10-07